One column at a time McBASC scores are now cached (in a directory called oneD in the results directory).  WriteScores has been updated to rely on these cached values.  This means that once McBASC 1D is calculated the entire PFAM dataset can be done in hours not days on an 8 core box. Should allow us to knock out the last few bugs more quickly.
Fixed a bug in HelixSheetGroup where if the same structure was repeated
twice in a PDB file, it was compared to itself.  Made it so that if two
structures bordered each other, the column in the middle (that is shared
by both) is assigned to the right structure. 

Added WriteTruncatedPfamAlignments that writes a version of PFAM-a that
has only the alignments that we are concerned with.  This yields a file
that is only 70MB, which is much easier to deal with and parse than the
full PFAM-A file.